Global Styles: Fetch the variations inside the component
