+++ packages/edit-site/src/components/global-styles/variations/variations-typography.js
@@ -35,6 +35,10 @@ export default function TypographyVariations() {
 
 	const { base } = useContext( GlobalStylesContext );
 
+	if ( ! typographyVariations.length ) {
+		return null;
+	}
+
 	/*
 	 * Filter duplicate variations based on the font families used in the variation.
 	 */
